ICS_scatter {ICS}	R Documentation

Location and Scatter Estimates for ICS

Description

Computes a scatter matrix and an optional location vector to be used in transforming the data to an invariant coordinate system or independent components.

Usage

ICS_cov(x, location = TRUE)

ICS_cov4(x, location = c("mean", "mean3", "none"))

ICS_covW(x, location = TRUE, alpha = 1, cf = 1)

ICS_covAxis(x, location = TRUE)

ICS_tM(x, location = TRUE, df = 1, ...)

ICS_scovq(x, y, ...)

Arguments

x
a numeric matrix or data frame.

location
for ICS_cov(), ICS_cov4(), ICS_covW(), and ICS_covAxis(), a logical indicating whether to include the sample mean as location estimate (default to TRUE). For ICS_cov4(), alternatively a character string specifying the location estimate can be supplied. Possible values are "mean" for the sample mean (the default), "mean3" for a location estimate based on third moments, or "none" to not include a location estimate. For ICS_tM() a logical inficating whether to include the M-estimate of location (default to TRUE).

alpha
parameter of the one-step M-estimator (default to 1).

cf
consistency factor of the one-step M-estimator (default to 1).

df
assumed degrees of freedom of the t-distribution (default to 1, which corresponds to the Cauchy distribution).

...
additional arguments to be passed down to scovq().

y
numerical vector specifying the dependent variable.

Details

ICS_cov() is a wrapper for the sample covariance matrix as computed by cov().

ICS_cov4() is a wrapper for the scatter matrix based on fourth moments as computed by cov4(). Note that the scatter matrix is always computed with respect to the sample mean, even though the returned location component can be specified to be based on third moments as computed by mean3(). Setting a location component other than the sample mean can be used to fix the signs of the invariant coordinates in ICS() based on generalized skewness values, for instance when using the scatter pair ICS_cov() and ICS_cov4().

ICS_covW() is a wrapper for the one-step M-estimator of scatter as computed by covW().

ICS_covAxis() is a wrapper for the one-step Tyler shape matrix as computed by covAxis(), which is can be used to perform Principal Axis Analysis.

ICS_tM() is a wrapper for the M-estimator of location and scatter for a multivariate t-distribution, as computed by tM().

ICS_scovq() is a wrapper for the supervised scatter matrix based on quantiles scatter, as computed by scovq().

Value

An object of class "ICS_scatter" with the following components:

location
if requested, a numeric vector giving the location estimate.

scatter
a numeric matrix giving the estimate of the scatter matrix.

label
a character string providing a label for the scatter matrix.

Author(s)

Andreas Alfons and Aurore Archimbaud

References

Arslan, O., Constable, P.D.L. and Kent, J.T. (1995) Convergence behaviour of the EM algorithm for the multivariate t-distribution, Communications in Statistics, Theory and Methods, 24(12), 2981–3000. doi: 10.1080/03610929508831664.

Critchley, F., Pires, A. and Amado, C. (2006) Principal Axis Analysis. Technical Report, 06/14. The Open University, Milton Keynes.

Kent, J.T., Tyler, D.E. and Vardi, Y. (1994) A curious likelihood identity for the multivariate t-distribution, Communications in Statistics, Simulation and Computation, 23(2), 441–453. doi: 10.1080/03610919408813180.

Oja, H., Sirkia, S. and Eriksson, J. (2006) Scatter Matrices and Independent Component Analysis. Austrian Journal of Statistics, 35(2&3), 175-189.

Tyler, D.E., Critchley, F., Duembgen, L. and Oja, H. (2009) Invariant Co-ordinate Selection. Journal of the Royal Statistical Society, Series B, 71(3), 549–592. doi: 10.1111/j.1467-9868.2009.00706.x.

See Also

ICS()

colMeans(), mean3()

cov(), cov4(), covW(), covAxis(), tM(), scovq()

Examples

data("iris")
X <- iris[,-5]
ICS_cov(X)
ICS_cov4(X)
ICS_covW(X, alpha = 1, cf = 1/(ncol(X)+2))
ICS_covAxis(X)
ICS_tM(X)

# The number of explaining variables
p <- 10
# The number of observations
n <- 400
# The error variance
sigma <- 0.5
# The explaining variables
X <- matrix(rnorm(p*n),n,p)
# The error term
epsilon <- rnorm(n, sd = sigma)
# The response
y <- X[,1]^2 + X[,2]^2*epsilon
ICS_scovq(X, y = y)

[Package ICS version 1.4-2 Index]