Mean3Cov4 {ICS}	R Documentation

Location Vector Based on 3rd Moments and Scatter Matrix Based on 4th Moments

Description

Returns, for some multivariate data, the location vector based on 3rd moments and the scatter matrix based on 4th moments.

Usage

Mean3Cov4(x)

Arguments

x
a numeric data matrix.

Details

Note that the scatter matrix of 4th moments is computed with respect to the mean vector and not with respect to the location vector based on 3rd moments.

Value

A list containing:

locations
The location vector based on 3rd moments as computed by mean3.

scatter
The scatter matrix based on 4th moments as computed by cov4.

Author(s)

Klaus Nordhausen

See Also

mean3, cov4

Examples

X <- rmvnorm(200, 1:3, diag(2:4))
Mean3Cov4(X)

[Package ICS version 1.4-2 Index]