coef.ICS-S3 {ICS}	R Documentation

To extract the Coefficient Matrix of the ICS Transformation

Description

Extracts the coefficient matrix of a linear transformation to an invariant coordinate system. Each row of the matrix contains the coefficients of the transformation to the corresponding component.

Usage

## S3 method for class 'ICS'
coef(object, select = NULL, drop = FALSE, index = NULL, ...)

Arguments

object
an object inheriting from class "ICS" containing results from an ICS transformation.

select
an integer, character, or logical vector specifying for which components to extract the coefficients, or NULL to extract the coefficients for all components.

drop
a logical indicating whether to return a vector rather than a matrix in case coefficients are extracted for a single component (default to FALSE).

index
an integer vector specifying for which components to extract the coefficients, or NULL to extract coefficients for all components. Note that index is deprecated and may be removed in the future, use select instead.

...
additional arguments are ignored.

Value

A numeric matrix or vector containing the coefficients for the requested components.

Author(s)

Andreas Alfons and Aurore Archimbaud

See Also

ICS()

gen_kurtosis(), components(), fitted(), and plot() methods

Examples

data("iris")
X <- iris[,-5]
out <- ICS(X)
coef(out)
coef(out, select = c(1,4))
coef(out, select = 1, drop = FALSE)

[Package ICS version 1.4-2 Index]