components {ICS}	R Documentation

To extract the Component Scores of the ICS Transformation

Description

Extracts the components scores of an invariant coordinate system obtained via an ICS transformation.

Usage

components(x, ...)

## S3 method for class 'ICS'
components(x, select = NULL, drop = FALSE, index = NULL, ...)

Arguments

x
an object inheriting from class "ICS" containing results from an ICS transformation.

...
additional arguments to be passed down.

select
an integer, character, or logical vector specifying which components to extract, or NULL to extract all components.

drop
a logical indicating whether to return a vector rather than a matrix in case a single component is extracted (default to FALSE).

index
an integer vector specifying which components to extract, or NULL to extract all components. Note that index is deprecated and may be removed in the future, use select instead.

Value

A numeric matrix or vector containing the requested components.

Author(s)

Andreas Alfons and Aurore Archimbaud

See Also

ICS()

gen_kurtosis(), coef(), fitted(), and plot() methods

Examples

data("iris")
X <- iris[,-5]
out <- ICS(X)
components(out)
components(out, select = c(1,4))
components(out, select = 1, drop = FALSE)

[Package ICS version 1.4-2 Index]