cov4 {ICS}	R Documentation

Scatter Matrix based on Fourth Moments

Description

Estimates the scatter matrix based on the 4th moments of the data.

Usage

cov4(X, location = "Mean", na.action = na.fail)

Arguments

X
numeric data matrix or dataframe, missing values are not allowed.

location
can be either Mean, Origin or numeric. If numeric the matrix is computed wrt to the given location.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

If location is Mean the scatter matrix of 4th moments is computed wrt to the sample mean. For location = Origin it is the scatter matrix of 4th moments wrt to the origin. The scatter matrix is standardized in such a way to be consistent for the regular covariance matrix at the multinormal model. It is given for n x p matrix X by

1/(p+2) ave{[(x_i-x_bar)S^{-1}(x_i-x_bar)'] (x_i-x_bar)'(x_i-x_bar)},

where x_bar is the mean vector and S the regular covariance matrix.

Value

A matrix containing the estimated fourth moments scatter.

Author(s)

Klaus Nordhausen

References

Cardoso, J.F. (1989), Source separation using higher order moments, in Proc. IEEE Conf. on Acoustics, Speech and Signal Processing (ICASSP'89), 2109–2112. <doi:10.1109/ICASSP.1989.266878>.

Oja, H., Sirki?, S. and Eriksson, J. (2006), Scatter matrices and independent component analysis, Austrian Journal of Statistics, 35, 175–189.

Examples

set.seed(654321)
cov.matrix <- matrix(c(3,2,1,2,4,-0.5,1,-0.5,2), ncol=3)
X <- rmvnorm(100, c(0,0,0), cov.matrix)
cov4(X)
cov4(X, location="Origin")
rm(.Random.seed)

[Package ICS version 1.4-2 Index]