covAxis {ICS}	R Documentation

One step Tyler Shape Matrix

Description

This matrix can be used to get the principal axes from ics, which is then known as principal axis analysis.

Usage

covAxis(X, na.action = na.fail)

Arguments

X
numeric data matrix or dataframe.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

The covAxis matrix V is a given for a n x p data matrix X as

p ave{[(x_i-x_bar)S^{-1}(x_i-x_bar)']^(-1) (x_i-x_bar)'(x_i-x_bar)},

where x_bar is the mean vector and S the regular covariance matrix.

covAxis can be used to perform a Prinzipal Axis Analysis (Critchley et al. 2006) using the function ics. In that case, for a centered data matrix X, covAxis can be used as S2 in ics, where S1 should be in that case the regular covariance matrix.

Value

A matrix containing the estimated one step Tyler shape matrix.

Author(s)

Klaus Nordhausen

References

Critchley , F., Pires, A. and Amado, C. (2006), Principal axis analysis, Technical Report, 06/14, The Open University Milton Keynes.

Tyler, D.E., Critchley, F., D?mbgen, L. and Oja, H. (2009), Invariant co-ordinate selecetion, Journal of the Royal Statistical Society,Series B, 71, 549–592. <doi:10.1111/j.1467-9868.2009.00706.x>.

See Also

ics

Examples

data(iris)
iris.centered <- sweep(iris[,1:4], 2, colMeans(iris[,1:4]), "-")
iris.paa <- ics(iris.centered, cov, covAxis, stdKurt = FALSE)
summary(iris.paa)
plot(iris.paa, col=as.numeric(iris[,5]))
mean(iris.paa@gKurt)
emp.align <- iris.paa@gKurt
emp.align

screeplot(iris.paa)
abline(h = 1)

[Package ICS version 1.4-2 Index]