covOrigin {ICS}	R Documentation

Covariance Matrix with Respect to the Origin

Description

Estimates the covariance matrix with respect to the origin.

Usage

covOrigin(X, location = NULL, na.action = na.fail)

Arguments

X
a numeric data matrix or dataframe.

location
optional location value which serves as the center instead of the origin.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

The covariance matrix S_0 with respect to origin is given for a matrix X with n observations by

S_0=X'X/n.

Value

A matrix containing the estimated covariance matrix with respect to the origin.

Author(s)

Klaus Nordhausen

See Also

cov

Examples

set.seed(654321)
cov.matrix <- matrix(c(3,2,1,2,4,-0.5,1,-0.5,2), ncol=3)
X <- rmvnorm(100,c(0,0,0),cov.matrix)
covOrigin(X)
rm(.Random.seed)

[Package ICS version 1.4-2 Index]