fitted.ICS-S3 {ICS}	R Documentation

Fitted Values of the ICS Transformation

Description

Computes the fitted values based on an invariant coordinate system obtained via an ICS transformation. When using all components, computing the fitted values constitutes a backtransformation to the observed data. When using fewer components, the fitted values can often be viewed as reconstructions of the observed data with noise removed.

Usage

## S3 method for class 'ICS'
fitted(object, select = NULL, index = NULL, ...)

Arguments

object
an object inheriting from class "ICS" containing results from an ICS transformation.

select
an integer, character, or logical vector specifying which components to use for computing the fitted values, or NULL to compute the fitted values from all components.

index
an integer vector specifying which components to use for computing the fitted values, or NULL to compute the fitted values from all components. Note that index is deprecated and may be removed in the future, use select instead.

...
additional arguments are ignored.

Value

A numeric matrix containing the fitted values.

Author(s)

Andreas Alfons and Aurore Archimbaud

See Also

ICS()

gen_kurtosis(), coef(), components(), and plot() methods

Examples

data("iris")
X <- iris[,-5]
out <- ICS(X)
fitted(out)
fitted(out, select = 4)

[Package ICS version 1.4-2 Index]