gen_kurtosis {ICS}	R Documentation

To extract the Generalized Kurtosis Values of the ICS Transformation

Description

Extracts the generalized kurtosis values of the components obtained via an ICS transformation.

Usage

gen_kurtosis(object, ...)

## S3 method for class 'ICS'
gen_kurtosis(object, select = NULL, scale = FALSE, index = NULL, ...)

Arguments

object
an object inheriting from class "ICS" containing results from an ICS transformation.

...
additional arguments to be passed down.

select
an integer, character, or logical vector specifying for which components to extract the generalized kurtosis values, or NULL to extract the generalized kurtosis values of all components.

scale
a logical indicating whether to scale the generalized kurtosis values to have product 1 (default to FALSE). See ‘Details’ for more information.

index
an integer vector specifying for which components to extract the generalized kurtosis values, or NULL to extract the generalized kurtosis values of all components. Note that index is deprecated and may be removed in the future, use select instead.

Details

The argument scale is useful when ICS is performed with shape matrices rather than true scatter matrices. Let S1 and S2 denote the scatter or shape matrices used in ICS.

If both S1 and S2 are true scatter matrices, their order in principal does not matter. Changing their order will just reverse the order of the components and invert the corresponding generalized kurtosis values.

The same does not hold when at least one of them is a shape matrix rather than a true scatter matrix. In that case, changing their order will also reverse the order of the components, but the ratio of the generalized kurtosis values is no longer 1 but only a constant. This is due to the fact that when shape matrices are used, the generalized kurtosis values are only relative ones. It is then useful to scale the generalized kurtosis values such that their product is 1.

Value

A numeric vector containing the generalized kurtosis values of the requested components.

Author(s)

Andreas Alfons and Aurore Archimbaud

See Also

ICS()

coef(), components(), fitted(), and plot() methods

Examples

data("iris")
X <- iris[,-5]
out <- ICS(X)
gen_kurtosis(out)
gen_kurtosis(out, scale = TRUE)
gen_kurtosis(out, select = c(1,4))

[Package ICS version 1.4-2 Index]