mean3 {ICS}	R Documentation

Location Estimate based on Third Moments

Description

Estimates the location based on third moments.

Usage

mean3(X, na.action = na.fail)

Arguments

X
numeric data matrix or dataframe with at least two columns.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

This location estimate is defined for a n x p matrix X as

1/p ave{[(x_i-x_bar)S^{-1}(x_i-x_bar)'] x_i},

where x_bar is the mean vector and S the regular covariance matrix.

Value

A vector.

Author(s)

Klaus Nordhausen

References

Oja, H., Sirki?, S. and Eriksson, J. (2006), Scatter matrices and independent component analysis, Austrian Journal of Statistics, 35, 175–189.

Examples

set.seed(654321)
cov.matrix <- matrix(c(3,2,1,2,4,-0.5,1,-0.5,2), ncol=3)
X <- rmvnorm(100, c(0,0,0), cov.matrix)
mean3(X)
rm(.Random.seed)

[Package ICS version 1.4-2 Index]