mvnorm.kur.test {ICS}	R Documentation

Test of Multivariate Normality Based on Kurtosis

Description

Test for multivariate normality which uses as criterion the kurtosis measured by the ratio of regular covariance matrix and matrix of fourth moments.

Usage

mvnorm.kur.test(X, method = "integration", n.simu = 1000, 
                na.action = na.fail)

Arguments

X
a numeric data frame or matrix.

method
defines the method used for the computation of the p-value. The possibilites are "integration" (default), "satterthwaite" or "simulation". Details below.

n.simu
if 'method=simulation' this specifies the number of replications in the simulation.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

This test implements the multivariate normality test based on kurtosis measured by two different scatter estimates as described in Kankainen, Taskinen and Oja. The choice here is based on the regular covariance matrix and matrix of fourth moments (cov4). The limiting distribution of the test statistic W is a linear combination of independent chi-square variables with different degrees of freedom. Exact limiting p-values or approximated p-values are obtained by using the function pchisqsum. However Kankainen et al. mention that even for n = 200 the convergence can be poor, therefore also p-values simulated under the NULL can be obtained.

Note that the test statistic used is a symmetric version of the one in the paper to guarantee affine invariance.

Value

A list with class 'htest' containing the following components:

statistic
the value of the test statistic W.

parameter
the degrees of freedom for the test statistic W with their weights or the number of replications depending on the chosen method.

p.value
the p-value for the test.

method
a character string indicating what type of test was performed.

data.name
a character string giving the name of the data.

Author(s)

Klaus Nordhausen

References

Kankainen, A., Taskinen, S. and Oja, H. (2007), Tests of multinormality based on location vectors and scatter matrices, Statistical Methods and Applications, 16, 357–379. <doi:10.1007/s10260-007-0045-9>.

See Also

mvnorm.skew.test

Examples

X<-rmvnorm(100, c(2, 4, 5))
mvnorm.kur.test(X)
mvnorm.kur.test(X, method = "satt")
mvnorm.kur.test(X, method = "simu")

[Package ICS version 1.4-2 Index]