mvnorm.skew.test {ICS}	R Documentation

Test of Multivariate Normality Based on Skewness

Description

Test for multivariate normality that uses as criterion the skewness measured as the difference between location estimates based on first respectively third moments

Usage

mvnorm.skew.test(X, na.action = na.fail)

Arguments

X
a numeric data frame or matrix.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

This test implements the multivariate normality test based on skewness measured by two different location estimates as described in Kankainen, Taskinen and Oja. The choice here is based on the regular mean vector and the location estimate based on third moments (mean3). The scatter matrix used is the regular covariance matrix.

Value

A list with class 'htest' containing the following components:

statistic
the value of the test statistic U.

parameter
the degrees of freedom for the statistic U.

p.value
the p-value for the test.

method
a character string indicating what type of test was performed.

data.name
a character string giving the name of the data.

Author(s)

Klaus Nordhausen

References

Kankainen, A., Taskinen, S. and Oja, H. (2007),Tests of multinormality based on location vectors and scatter matrices, Statistical Methods and Applications, 16, 357–379. <doi:10.1007/s10260-007-0045-9>.

See Also

mvnorm.kur.test

Examples

X<-rmvnorm(100,c(2,4,5))
mvnorm.skew.test(X)

[Package ICS version 1.4-2 Index]