plot.ICS-S3 {ICS}	R Documentation

Scatterplot Matrix of Component Scores from the ICS Transformation

Description

Produces a scatterplot matrix of the component scores of an invariant coordinate system obtained via an ICS transformation.

Usage

## S3 method for class 'ICS'
plot(x, select = NULL, index = NULL, ...)

Arguments

x
an object inheriting from class "ICS" containing results from an ICS transformation.

select
an integer, character, or logical vector specifying which components to plot. If NULL, all components are plotted if there are at most six components, otherwise the first three and the last three components are plotted (as the components with extreme generalized kurtosis values are the most interesting ones).

index
an integer vector specifying which components to plot, or NULL to plot all components. Note that index is deprecated and may be removed in the future, use select instead.

...
additional arguments to be passed down to pairs().

Author(s)

Andreas Alfons and Aurore Archimbaud

See Also

ICS()

gen_kurtosis(), coef(), components(), and fitted() methods

Examples

data("iris")
X <- iris[,-5]
out <- ICS(X)
plot(out)
plot(out, select = c(1,4))

[Package ICS version 1.4-2 Index]