print.ICS-S3 {ICS}	R Documentation

Basic information of ICS Object

Description

Prints information of an ICS object.

Usage

## S3 method for class 'ICS'
print(x, info = FALSE, digits = 4L, ...)

Arguments

x
object of class ICS.

info
Logical, either TRUE or FALSE. If TRUE, print additional information on arguments used for computing scatter matrices (only named arguments that contain numeric, character, or logical scalars) and information on the parameters of the algorithm. Default is FALSE.

digits
number of digits for the numeric output.

...
additional arguments passed to print()

Author(s)

Andreas Alfons and Aurore Archimbaud

See Also

ICS()

Examples

data("iris")
X <- iris[,-5]
out <- ICS(X)
print(out)
print(out, info = TRUE)

[Package ICS version 1.4-2 Index]