screeplot.ics {ICS}	R Documentation

Screeplot for an ICS Object

Description

Plots the kurtosis measures of an ics object against its index number. Two versions of this screeplot are available.

Usage

## S3 method for class 'ics'
screeplot(x, index = NULL, type = "barplot", 
          main = deparse(substitute(x)), ylab = "generalized kurtosis", 
          xlab = "component", names.arg = index, labels = TRUE, ...)

Arguments

x
object of class ics.

index
index of the components to be plottes. If NULL all components are used.

type
barplot if a barplot or lines if a line plot is preferred.

main
main title of the plot.

ylab
y-axis label.

xlab
x-axis label.

names.arg
names.arg argument passed on to barplot.

labels
labels argument for the labels of the x-axis passed on to axis.

...
other arguments for the plotting functions.

Author(s)

Klaus Nordhausen

See Also

plot.ics, ics-class and ics

Examples

set.seed(654321)
A <- matrix(c(3,2,1,2,4,-0.5,1,-0.5,2),ncol=3)
eigen.A <- eigen(A)
sqrt.A <- eigen.A$vectors %*% (diag(eigen.A$values))^0.5 %*% t(eigen.A$vectors)
normal.ic <- cbind(rnorm(800), rnorm(800), rnorm(800))
mix.ic <- cbind(rt(800,4), rnorm(800), runif(800,-2,2))

data.normal <- normal.ic %*% t(sqrt.A)
data.mix <- mix.ic %*% t(sqrt.A)

par(mfrow=c(1,2))
screeplot(ics(data.normal))
screeplot(ics(data.mix), type="lines")
par(mfrow=c(1,1))
rm(.Random.seed)

screeplot(ics(data.normal), names.arg=paste("IC", 1:ncol(A), sep=""), xlab="")

[Package ICS version 1.4-2 Index]